backend/requirements.txt
web3==6.15.1
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
pandas==2.1.4
